JVMCBC-743: Remove Authenticator from Environment
This changeset removes the Authenticator from the Environment and instead
puts it into the core context. All the other code changes are refactrings
of using/setting the authenticator in the right place rather than on the
environment.

NOTE: this breaks the build because scala needs to be done on top of this!

Change-Id: I37ca4984dbeca51d20e03e26fea09b1957dab078
Reviewed-on: http://review.couchbase.org/114953
Reviewed-by: Graham Pople <grahampople@gmail.com>
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>