JCBC-1408: Add AccessDeleted flag to Sub-Document lookupIn Add delWithMeta to core
This change adds the AccessDeleted flag to lookupIn, as required by the
RFC.  This flag allows data (xattrs only) to be returned from a
deleted document (a tombstone).

As a bonus, support for the delWithMeta command is added to core.  This
may or may not be needed for transactions so it's useful to get it in
now, plus it's used for the AccessDeleted testing.

DelWithMeta is deliberately not exposed through java-client, it's not
intended to be public API.

Change-Id: Ic7ac22ef65619fd88081b151031a492060446a7b
Reviewed-on: http://review.couchbase.org/115124
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Tested-by: Graham Pople <grahampople@gmail.com>