JCBC-1399: cluster.diagnostics function missing
Ports the SDK2 version of diagnostics across to SDK3.  This is a
starting point that sets us up with something for beta release, even
if we need to make changes later if the RFC changes.

There are a couple of minor differences:

1. The diagnostics RFC specifies a scope for EndpointHealth, which can
be used for KV to indicate which bucket it's for.  It wasn't exposed in
SDK2 for some reason, I added it here.  For the BUCKET_GLOBAL KV service,
it will be Optional.empty().

2. SDK2 wired things up by sending a diagnostics request packet down into
core, and then handling the Observable response.  I'm not sure all the
rigmarole is justified, since all this data is available to be returned
immediately.  I've made it simple non-async methods instead.

3. Changed the EndpointHealth output to a) redact host, and b) use
getHostString to return the hostname if it was resolved, else the
original string.  Which I think is what's wanted?

Change-Id: Id1f7fb739fe9e6d3eca9eadd06c60560dc2ed864
Reviewed-on: http://review.couchbase.org/115218
Reviewed-by: Graham Pople <grahampople@gmail.com>
Tested-by: Graham Pople <grahampople@gmail.com>