JCBC-1454 Datastructures need to lazy create.
Motivation
==========
RFC was explicit about lazily creating documents - so lets not create
the empty document (json array or json object) until we have something
to go in.  Also, insure that clear() actually removes the document.

Modification
===========
Removed the empty creation of the doc at construction, handled the
KeyNotFoundExceptions, and so forth.  While at it, I also noticed
I'd not properly extended the CouchbaseQueue's iterator, so did that
too.

Change-Id: Ib4aa8c15b4cdfe4a5d7969585d7f4d3e787b9bc5
Reviewed-on: http://review.couchbase.org/115273
Reviewed-by: Graham Pople <grahampople@gmail.com>
Tested-by: Build Bot <build@couchbase.com>