JCBC-1463: Move kv transcoder/serializers to option blocks
Motivation
----------
To align with the RFC and for consistency reasons, all custom serializers
and transcoders should be put at the option level. This changeset aligns

 - get, getandlock, getandtouch
 - subdoc lookup in, subdoc mutate in

to these guidelines.

Note that the mutate in command serializers have been deprecated so we
do not break txns for now, this will be removed in beta refresh.

Change-Id: I9860f9df1f5fe76a693d9fc3333a0886d688e037
Reviewed-on: http://review.couchbase.org/115319
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>