JCBC-1466: Remove redundant DataFormat
Motivation
----------
This changeset builds on the previous big refactoring and further
discussion around the RFC semantics for transcoding. After careful
consideration of the different options and backwards compatibility
aspects it has been found that:

 - The DataFormat just duplicates the logic we can apply with custom
   transcoders in the first place.
 - Different transcoders are sufficient to handle all scenarios we need
   to handle with common flags and also custom formats.

Modifications
-------------
As a result, this changeset removes the DataFormat enum and its associates
code and tests. This also makes it necessary to split up the default transcoder
into a couple of special-purpose transcoders where the Json one is still the
default.

To avoid ambiguity, the default json transcoder does NOT accept a byte array as
input or output. That's what the RawJsonTranscoder (or the other raw ones with
different flags) are for, and the transcoder will point this out to the user so
it will be caught at dev time.

Result
------
It is now way eaiser to use custom data formats, or serialized JSON: just use
a different transcoder and it will "just work".

Change-Id: Ib5a5efa865bab0430dfaf9bcf459bda2e81b3ff7
Reviewed-on: http://review.couchbase.org/115383
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Graham Pople <grahampople@gmail.com>