JVMCBC-748: Seed hosts should only include KV nodes on update
Motivation
----------
Previously we added a feature which keeps the seed nodes up-to-date
as the configuration state changes. Unfortunately it added all the
hostnames, but it should only add the nodes which have the KV
service enabled in the first place.

This avoids the situation where subsequent bucket open attempts
will reach nodes that do not have KV enabled and as a result have
"ugly" warnings in the logs - although not considered harmful at
this point.

Change-Id: I1dd546603d50f008919cbe974f741fa06cfad163
Reviewed-on: http://review.couchbase.org/115558
Reviewed-by: Graham Pople <grahampople@gmail.com>
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>