SCBC-99: Remove timeout on blocking
(This is an implementation detail that should be a no-op.)

During core's early days it had timeout support, so the Scala client,
whenever it needed to block on a result from core, added a safety
timeout.

Now core has timeouts, rendering these redundant.  All blocks are
converted to infinite ones.
Separate block and blockForever methods aren't needed, so blockForever
is renamed.

Change-Id: I5f5f55ffdfff1839be62b4f908cc167b6a72978c
Reviewed-on: http://review.couchbase.org/115611
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Graham Pople <grahampople@gmail.com>