JCBC-1473 CouchbaseArrayListTest fails when not run against the mock
Motivation
==========
Tests should pass on both mock and real CB server.

Modification
============
We were making a MutateIn call with options which specified both a cas and an upsert
store semantics.  This results in a result code of 0x04, which we don't handle and means
the request was invalid, in cases where the doc doesn't exist.  But the mock gives a NOT
FOUND which becomes a KeyNotFoundException we handled.

Since this happens only in the iterator's add call, if we happen to be just adding to an empty
doc, we _do_ want to create an empty doc and add this element.  So - we check to see if the
delegate iterator is empty, and if so we can just create an empty list and try again.  Added
tests for the various possibilities.

Change-Id: Id601cb6798919f00726c768b3a06c5f712a2ecd9
Reviewed-on: http://review.couchbase.org/115630
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Graham Pople <grahampople@gmail.com>