SCBC-62: Add search index manager API
This provides the minimum CRUD index management to support creating
search indexes for test scaffolding, e.g. it focuses on the 90% use-case
that users are most likely to require.

Support for the more esoteric search index management APIs
(such as freeze plans) can come later, if required.

As with other management APIs, scaladocs will come later in a future
sweep.

Change-Id: I198b3236288fca9c94b3216fedf8aa3c4cba38c8
Reviewed-on: http://review.couchbase.org/116370
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Graham Pople <grahampople@gmail.com>