JVMCBC-750: Improve accuracy of timer
Motivation
----------
We are scheduling operations for retry also through the timer and not
just track timeouts, so the default accuracy of 100ms is a little bit
too much (i.e. if a 1ms delay is scheduled).

In testing we found that 10ms strikes a good balance between accuracy
and noisiness so we are going with that for now.

Change-Id: I3dc968896b45b18a98b9552c7cfbcf87143e5358
Reviewed-on: http://review.couchbase.org/116498
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Graham Pople <grahampople@gmail.com>