JCBC-1477: View query does never complete when 0 rows returned
Motivation
----------
During testing it has been observed that view queries return but
never complete when no rows are returned as part of the view
result.

Modifications
-------------
This bug has been around since we intrduced support for the "debug"
property of view queries in SDK 3. The root cause of this was tracked
down to a logic bug where the header has been only marked as ready
when either a debug header is present OR the rows start to stream.

This meant that if debug was not enabled and no rows were returned,
the header was never marked as being present and as a result the
ViewResponse from core never got generated in the first place.

Based on the view streaming response observation that debug information
is always returned before the total_rows, we can assume that the header
is complete once we got the total_rows which are always present. This
simplifies the parsing logic and also fixes the issue at hand.

An integration test has been added that verifies 0 and non-0 row
responses. Note that is enabled for the mock as well because the
mock supports view queries.

Change-Id: I12c1ce90f89f88686731a4d4e0b300be66cd6311
Reviewed-on: http://review.couchbase.org/116561
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Graham Pople <grahampople@gmail.com>
Reviewed-by: David Kelly <davidmichaelkelly@gmail.com>