JVMCBC-753: Add completionCallback to Circuit Breaker
Motivation
----------
Aligning with the RFC, the circuit breaker should contain a
completionCallback which allows the user to customize the
type of exceptions to accept.

Modifications
-------------
Alongside adding the callback, the builder is upgraded to fit
with the general building strategy of the SDK, it now uses
only cancellations (which include timeouts) to signal a failed
request to the circuit breaker by default.

Tests have been added and modified to assert this behavior.

Change-Id: I39d0ca5e1e0561779c4ad7c2f8708b16832dff8f
Reviewed-on: http://review.couchbase.org/116658
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>