JVMCBC-637: Add "logical complete" time to Request Context.
As a preparation for tracing, it is important that the SDK is able to track
a "logical" complete timestamp, which happens when all the logical sub-parts
are also complete. Most of the time this is very close to just completing the
future, but in the case of observe-based operations it will only happen after
the observe is done.

For now only certain mutate-ops are tracked, this will be expanded in a subsequent
commit once the approach is fully vetted.

Change-Id: I7bd6d91bdb341c0de995bb1eb655e1debe8ec632
Reviewed-on: http://review.couchbase.org/116843
Reviewed-by: Graham Pople <grahampople@gmail.com>
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>