Adding scalafmt so Scala code is automatically formatted.
This is driven by the .scalafmt.conf configuration file, which
contains a basic config.  I'm fairly happy with the results, we can
always adjust it later.

mvn will now automatically reformat the code at the validate stage.

Instructions added to README on how to configure IntelliJ to use
scalafmt for formatting.

This change doesn't prevent unformatted code getting through gerrit.
I may later add either a) a git hook to do `mvn validate`, though it's
a little slow, or b) force validation on gerrit, though I'm not certain
that's possible with this plugin.

The reformatted sources will follow in a later changeset.

Change-Id: I3ffeefbe0b01e8b9ddb8e0e185bb269e4b4e9dc3
Reviewed-on: http://review.couchbase.org/116849
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>