JVMCBC-637: Refactor to and from dispatch tags in RequestContext
Motivation
----------
To properly fill in the details for RTO, each request needs to know to
wich socket it got dispatched, both remote and local. To avoid recreating
this info over and over again, it is cached in the endpoint context.

Modifications
-------------
This changeset modifies the RequestContext to hold both the local and
remote side of the socket. Since Java does not have a trivial tuple
for unresolved hostname and port, this changeset adds one and modifies
all the interacting parties to hold this class instead of strings.

Note that the ping command added something similar but only focussed on
the ping command, so this changeset gets rid of that special logic and
just makes it use the generic functionality right now.

Hooking it up with RTO will come in a follow-up commit.

Change-Id: If3d6072d4f841a758ef1513ea6f1023f0b5260fc
Reviewed-on: http://review.couchbase.org/116856
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>