JVMCBC-637: Add basic support for RTO
Motivation
----------
SDK 3 needs to support our RTO feature like we did in SDK 2. This changeset
is the first of a couple that implements RTO.

Note that this change is not complete in the sense that it is not covering
all operations - this is done in further commits to not complicate this
changeset further and make review harder.

Modifications
-------------
Overall, this changeset:

 - Introduces the RequestTracer and RequestSpan interfaces
 - Provides a stub Noop implementation
 - Implements the default ThresholdRequestTracer ("RTO")
 - Adds new modules to the monorepo that provides support for OpenTracing as
   well as preliminary support for OpenCensus based on their snapshot repo

By having our own interfaces we do not have to bind ourselves to opentracing
or any other similar library. Support for those is provided with additional
modules which can be plugged in easily and used as needed.

Change-Id: Icd3f1674918b8649e7392fc5b08625aead9dae34
Reviewed-on: http://review.couchbase.org/116922
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Graham Pople <grahampople@gmail.com>
Reviewed-by: David Nault <david.nault@couchbase.com>