SCBC-98: Update JSON serialization
This completely reworks the JSON serialization, to make it more in-line
with the current RFC.

Changes Made
------------
Encodable renamed to JsonSerializer and aligned with RFC.
Decodable renamed to JsonDeserializer and aligned with RFC.
Transcoders added.
Many methods now take transcoders.
GetResult, LookupInResult, etc., now get passed those transcoders.
Any time transcoding takes place, the compiler-provided JsonSerializer/
JsonDeserializer is provided to the transcoder specified in the options,
as long as that transcoder can accept a serializer.

RFC Divergences
---------------
Scala provides the very useful capability to find an appropriate JSON
serializer for the required type at compile-time.  Taking advantage of
this means we can avoids the user
having to specify the requested type twice when using a non-standard
JSON library - e.g. once when specifying a custom serializer for the
transcoder, and once when doing a contentAs[T].

So, Scala diverges from the RFC in these ways:

1. JsonTranscoder doesn't take a serializer.  Instead, a JsonDeserializer[T]
is found at compile-time for contentAs[T].  And if a transcoder has been
provided in the options that can take a serializer (e.g. JsonTranscoder),
the found serializer is provided to the transcoder at the point of
deserialization.

2. Due to contravariant/covariant type issues, this meant splitting
the RFC's JsonSerializer into JsonSerializer and JsonDeserializer.

3. As a corollary, some methods that take serializers in the RFC (such as
QueryOptions), instead get those serializers found by the compiler at the
rowsAs[T] stage.

Change-Id: I0c7913192e5fd79ab052d3648a76f62f2d18f62d
Reviewed-on: http://review.couchbase.org/116957
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>