JVMCBC-752: Make Throwable the cause of an Event
Motivation
----------
This changeset makes the "cause" of an event (a throwable) a first
class concept on the interface. Once it is set, we can use it in the
logger and other consumers directly. The logger uses it now and passes
it to the imlementation, leading to much better debug output of exceptions
in the log, making it easier to debug issues.

Also, the events have been audited which right now already stored a
throwable and exposed them across the new unified interface.

Change-Id: I1b3a2984cdb0e6e01cbc53b7b0bb5cc29c8714f3
Reviewed-on: http://review.couchbase.org/117019
Reviewed-by: David Kelly <davidmichaelkelly@gmail.com>
Tested-by: Build Bot <build@couchbase.com>