core: fix flaky ManagerMessageHandlerTest
The root cause of the flakiness was that the response used a
DirectProcessor flux which just emitted items as they came along
and it could be that the subscriber in the test comes too late
seeing the config events.

The response now switches to a much better model: it still streams,
but keeps the latest one around cached. This also helps with late
subscription in the core itself, making it more robust. The test
is adapted to check that "at least one" of the two configs is seen
as emitted, since the cache only retains the latest one now.

Change-Id: I27b95a79c14f73673a16641146bcfdaac1b1f690
Reviewed-on: http://review.couchbase.org/117188
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Graham Pople <grahampople@gmail.com>