JVMCBC-758: Add upper cap for retry duration
The motivation behind this change is that it does not make sense to
schedule a retry timeout for longer than the actual complete
operation timeout (so if only 500ms left until an operation times
out, it does not make sense to queue it any longer than that).

This changeset introduces a cap calculation and applies it if
over the limit. Note that safeguard is put in place should it be
negative for some reason it returns the uncapped one.

Also the orchestrator has been modified that when a timer fires
to reschedule, we always check if it hasn't timed out already
before sending it back in.

Change-Id: Iadb1beb861865074f4b85e4a73039c340faccffd
Reviewed-on: http://review.couchbase.org/117608
Reviewed-by: Graham Pople <grahampople@gmail.com>
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>