JCBC-1498: Rework Error Handling 1/n
Motivation
----------
This changeset builds the foundation for the error handling
rework, and many more changes to come once the overall approach
is worked out in this one.

Modifications
-------------
This changeset sets the stage:

 - The CouchbaseException is extended with an ErrorContext which must
   be populated with specific contexts
 - The KeyValueErrorContext is introduced, in a full and reduced version
   depending on the scope thrown.
 - Retryable interface is removed, since it is not needed anymore and an
   old artifact of experiments.
 - Exceptions are renamed, added, removed as needed to align with the RFC
   error list spec.
 - The Get and GetAnd(lock/touch) methods are the guinea pigs for applying the
   new exception handling including tests where it makes sense. Javadoc
   applied to the blocking API for now, but will eventually be copied
   over to async and reactive as well.

Also, if you are going to spot some code in the Reactor class which looks
a bit nasty (but fear not!): previously when a mono is cancelled we'd
propagate it into the completable future which then in turn would cancel
the mono again. To avoid this, reactor code has some logic in place to
deal with this side effect on cancellation. Unfortunately with changing
our exception hierachy we cannot rely on extending the CancellationException
anymore so reactors code doesn't handle it. Checked with the reactor team
and we now have our own "fromFuture" implementation which is exactly
the same as in upstream but handles our STOPPED_LISTENING case explicitly
so to not trigger the drop hook.

NOTE: scala has been modified to work with stubs and needs to be modified
on its own to work with the new error handling infrastructure as the java
one does now.

Change-Id: Iab1648687587df6c83bde31290b72916d8134972
Reviewed-on: http://review.couchbase.org/117635
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Graham Pople <grahampople@gmail.com>