JCBC-1504: Cluster#connect needs to defer IO-related errors.
This changeset is a first step towards making the whole connect
pipeline "lazy", in the sense that all I/O is deferred until the
actual request is made.

In particular:

 - Async and Reactive variants do not return CompletableFuture or
   Mono for connect
 - The call to perform "gcccp" does not block the creation and instead
   is done in the background
 - Since now cluster-level calls can stumble on gcccp-in-progress do
   not cancel them immediately while it is still going on.

Note that unrelated but important, mock tests were failing because the
custom HTTP port was not set anymore. This changeset also allows to
explicitly set SeedNodes (marked as volatile) on the options and we
utilize it in the tests. So now the tests against the mock pass
properly again.

Change-Id: Ibf9ba3b49921f9266e8dadd9a609778f9cee3464
Reviewed-on: http://review.couchbase.org/117808
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Graham Pople <grahampople@gmail.com>