JCBC-1510: Opening a bucket and collections  must be lazy.
Similar to Cluster#connect, opening a bucket must
defer all errors into the operation. This changeset
does exactly that:

 - CompletableFuture<> and Mono<> have been removed
   and the AsyncBucket and ReactiveBucket are returned
   directly.
 - Same for Collection and Scopes
 - The internal core API has been reworked to not return
   a Mono but is triggered and asynchronously performs
   all the open bucket work.
 - The KV locator needs to take "config == null" into
   account, because a bucket might be currently in the
   process of being opened

Some tests had to be adapted as well. For now I had to remove
a couple unit tests which are not valid anymore, those checks
will be re-added on the actual kv operations later since
they are now deferred as well.

NOTE: this change only covers the "good" path, and likely
future changes are needed to propagate all kinds of errors
downstream and present them the right way.

Change-Id: Id7a99f0cda4009c8ccef7827b23e7c240cd1a482
Reviewed-on: http://review.couchbase.org/117815
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Graham Pople <grahampople@gmail.com>