JVMCBC-760: Add support for enabling security through connstr & properties
Motivation
----------
A user should be able to enable security through the connection string
and/or system properties, and this changeset makes sure they can.

For the connection string the scheme is used to enable it and both will
then allow for applying the trust certificate path as well. To make this
work a new overload has been added to the security config which is
convenient even on its own and allows to supply a trust certificate
path.

Note that it also adds a compat map to the connstr property loader so
it now also recognizes the lcb variant and we can add more in the future.

Regression tests added.

Change-Id: I0107ea9fe473707d70326c7f15200da1b0158f69
Reviewed-on: http://review.couchbase.org/117859
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Graham Pople <grahampople@gmail.com>