JVMCBC-762: Align service configuration with RFC
It has been decided that there are not going to be individual service
configurations but rather 3 simple settings across all SDKs:

	- numKvConnections (for the kv service)
	- maxHttpConnections (for all services but kv)
	- idleHttpConnectionTimeout

This makes it easy for users to navigate between SDKs and it also
simplifies the configuration aspect.

Since this removes the need for the ServiceConfig in java, it
has been removed. The internal components still use them, so the
Node just creates them on the fly from these params.

Note that by intention we are loosing the "min http sockets" per
service, but it has been decided that slow-start is seldomly a
problem for http based services so we can remove it for now and
if we really need it add it later easily.

Change-Id: I427cc21b9097cbc45e2b80780e6a88c605c6225e
Reviewed-on: http://review.couchbase.org/117885
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>