JCBC-1498: Rework error handling for "remove"
Note that this change just touches a lot of exceptions and this is
why it's so noisy.

Note that it also exposes the sync durability in the kv service
context if present, making it automatically dumped with the context
to aid debugging.

Also, the observe durability is not covered here since it would've
been even noisier and is coming in a separate changeset once
all mutation commands are converted.

Change-Id: I50e9ccb4cef3683f3fbb6b1cb1e3d17d8ee1c2bd
Reviewed-on: http://review.couchbase.org/117907
Reviewed-by: Graham Pople <grahampople@gmail.com>
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>