JCBC-1498: Rework error handling for "query"
This changeset reworks the error handling for the Query API based on the
current error handling RFC state.

Note that a couple of other exceptions had to be touched as well, especially
the authentication exception required very slight changes to the KV layer.

Also note that the QueryException is still there but needs to go away later.
I tried to remove it now, but is currently heavily used inside the query
management API, so we'll only be able to rip it out once all the error handling
for the query management API is also complete.

Change-Id: Ieba2de72fab24761cfd2709067125c78abefdf89
Reviewed-on: http://review.couchbase.org/118028
Reviewed-by: Graham Pople <grahampople@gmail.com>
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>