JVMCBC-763: Take bucket loading into account.
Because of the lazy loading, not only the cluster connect with
gcccp fetching needs to be taken into account, but also the
bucket fetching races with executing operations. This is especially
important with pre 6.5 clusters where the bucket open phase is
the first one where we'll touch the network.

This changeset also adds two specific error codes so it becomes clear
that operations were being retried during the bootstrap phase.

Change-Id: I3f874bb77d8b141a1730b3c0ff7bb6300b13628b
Reviewed-on: http://review.couchbase.org/118273
Reviewed-by: Graham Pople <grahampople@gmail.com>
Tested-by: Build Bot <build@couchbase.com>