JVMCBC-764: Unbreak encrypted connections
During bootstrap, the node identifier will have a port of 18091 because
that's the only thing we'll know. Over the course of opening and reconfig,
the node identifier will switch back to 8091 as it should be, coming
from the config.

This changeset retries the op until the config is in order and can be
dispatched properly. Please note that we *might* be able to infer
8091 == 8091 with being more intelligent on if tls is enabled when
verifying, but we can look at that when we find that the current
solution is not acceptable.

Change-Id: I31426c9913e1e9a3c1dbedd82ebacfbc4a5ec49f
Reviewed-on: http://review.couchbase.org/118297
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>