core: be more flexible with views not being found
During integration testing it has been detected that the server
might respond with different replies for a view that is not
found - this changeset covers more of these cases and also adds
more diagnostics to the CouchbaseException so it's easier to debug.

Also, one integration test is made more resilient to occasional
failures.

Change-Id: I6fe27523e70b9b8753a74333545a953c7bb15c29
Reviewed-on: http://review.couchbase.org/118366
Reviewed-by: Graham Pople <grahampople@gmail.com>
Tested-by: Build Bot <build@couchbase.com>