JCBC-1522: Instrument GetAllReplicas, GetAnyReplica
This changeset adds instrumentation for get all replicas and
get any replica (which in turn uses get all replicas under the
cover).

To make this work seamlessly with external tracers, the tracing
APIs had to be modified a bit to allow creating and finishing
request spans from the tracer wrapper implementation.

Change-Id: I409436535a763dfe4127eed9f21c0a6e97ad6cf9
Reviewed-on: http://review.couchbase.org/118701
Reviewed-by: Graham Pople <grahampople@gmail.com>
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>