JVMCBC-779: Allow to customize the event loop thread count.
We did allow customizing the "io pool size" in SDK 2, so we should
also allow overriding it in SDK 3. Usually it won't be needed, but
having the option to customize it without having to create a
event loop group manually makes it easier (and also configurable
via system properties potentially).

Change-Id: If76c8bff0c565a95b1a22ed020b97afa94060d65
Reviewed-on: http://review.couchbase.org/118863
Reviewed-by: Graham Pople <grahampople@gmail.com>
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>