Remove experimental diagnostics feature
This feature is removed from the stable 3.0 release because we are
intending to bring it back in a full and much better scope in one
of the following releases as "metrics". The code that is removed
now it is not lost though, it will serve as the basis for the
metrics feature.

Change-Id: I37176d6a414a9eccc7b2a76beb787579411fbf51
Reviewed-on: http://review.couchbase.org/119068
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Graham Pople <grahampople@gmail.com>