JCBC-1541: Remove explicit SubDocumentException
The SubDocumentException is not needed anymore, because the index
of the error is now part of the subdoc error context. This flattens
the structure and removes unnecessary duplication.

Test have been modified to not look for it. It also aligns it
with the latest RFC discussions (renaming CannotInsertValue to
ValueInvalid).

Change-Id: Iaaf80a91a9e6c009b834815f1e814b9822a17f5a
Reviewed-on: http://review.couchbase.org/119199
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Graham Pople <grahampople@gmail.com>