JVMCBC-781: Do not raise onErrorDropped on RequestCanceledException if client has stopped listening
If a Mono subscriber stops listening, it will cause the request to be
cancelled.  This should not lead to Reactor's onErrorDropped hook
being called - by default, this prints out the full stacktrace, and is
very spammy.

Transactions testing found that this scenario was leading to
onErrorDropped being called.  Working through the internals, it was
found that sometimes the RequestCanceledException is wrapped inside
a CompletionException.

This changeset adds unwrapping logic to handle this, and tests to handle
both scenarios.

Change-Id: I33fc1cf646a48c4595b74d616374eb3576d6dd5b
Reviewed-on: http://review.couchbase.org/119208
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Tested-by: Graham Pople <grahampople@gmail.com>
Tested-by: Build Bot <build@couchbase.com>