JVMCBC-782: Implement Orphan Reporter
This changeset brings forward the orphan reporter that has been
present in Java 2.x and makes use of it in the various message
handlers.

Every time a message handler wants to complete an operation but it
is already completed, it is sent to the orphan reporter instead and
will subsequently be logged at the configured interval.

Change-Id: I07e94f1ac719291e9e15d2f1d0522ceded3a7d39
Reviewed-on: http://review.couchbase.org/119376
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: David Nault <david.nault@couchbase.com>