JCBC-1545: Rework and polish the diagnostics API
This changeset updates and polishes the Diagnostics API by on the
one hand aligning it with the latest RFC iteration, but also by
polishing the javadocs, method signatures and implementation details
where needed.

It also introduces ping as an option and waitUntil ready that makes
use of both. There is still more polishing to be done, but this change
also modifies the test suite to use it instead of the "hack" and looks
good so far.

Change-Id: I7439fe170a288b534f41d5c9b5f7a09df66b7c71
Reviewed-on: http://review.couchbase.org/119409
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>