JVMCBC-784: Update http idle timeout to 30s.
The search service will terminate idle connections after 1 minute, and to
avoid this the default idle timeout is now 30s. This should still be plenty
to handle traffic surges, but avoid the search terminate issue.

Also, an event is now raised if a user configures a idle time over 1 minute.

Change-Id: Ie8aab734c1b228b12eba3694a86f8a274e08d08d
Reviewed-on: http://review.couchbase.org/119536
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>