JVMCBC-789: Gracefully handle collection loading in progress.
Since bootstrap is lazy, the collection map loading is too. Similar
to cluster and bucket-level configs loading in progress, the code
needs to handle the initial collection map loading. We latch onto
the exception being raised on encode to not penalize the hot code
path when successfully loaded and in-use.

Change-Id: I97b49dab0785061efe49e915f22dafda77c8edc4
Reviewed-on: http://review.couchbase.org/119813
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>