JCBC-1551: Remove "empty" constructor from JsonArray and JsonObject
The empty method has always been a needless duplicate of the create()
method, and before 3.0 GA let's take the opportunity to get rid of it.
Added instead is a create() method with an initial size overload so that
it can be constructed in a more efficient manner. Also some other minor
cleanups in the classes and tests.

Change-Id: Ie138d107cc165e60ffb98371ca27cc0cb4f3f66b
Reviewed-on: http://review.couchbase.org/119814
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>