JVMCBC-790: Reintroduce Backpressure
This changeset reintroduces the concept of a hard high watermark
backpressure into SDK 3. We had this concept in SDK 2 where it
was built-in implicitly into the request ringbuffer, but because
that one is gone now and we still need the concept for reliability
it is introduced in a different spot.

The main shortcoming of the old approach was that bad requests for
a subset of the nodes would fill up the buffer and requests coming
in that would go to healthy nodes were also penalized. The new approach
adds the backpressure mechanism to the timer API right before operations
are scheduled for retry. This has the effect that "good" operations
are not affected while still keeping an effective high watermark for
operations that go into the retry cycle. Since in overload conditions
operations will be rescheduled into the timer, they eventually would
fill up to the high watermark (default at 32k, so reasonably conservative)
and will then be cancelled with a descriptive cancellation reason.

In SDK 2 it was just called BackpressureException - now it is a RequestCanceledException
and in the context it shows up as TOO_MANY_REQUESTS_IN_RETRY.

Change-Id: Ib6536d8cd33a16f35512a0b7ddc828e7bb599f39
Reviewed-on: http://review.couchbase.org/119856
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>