gardening: rename free() on endpoint for more clarity
The free method name on the interface has been ambiguous because
it was used to signal both "free to write" but also "no outstanding
requests", which didn't really match up. It does not cause issues
right now because pipelined requests are not cleaned up on idle but
might be in the future.

This changeset clarifies the naming and avoids potential bugs in
the future with it.

Change-Id: Ie0cc5d72bf0226f739e055ffd00f495daf25f541
Reviewed-on: http://review.couchbase.org/119898
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Graham Pople <grahampople@gmail.com>