JVMCBC-767: Implement ConfigIdleRedialTimeout
This changeset adds the configuration option to the IoConfig and
then makes use of it in the manager message handler. The idea
is simple:

Every time a http streaming request is sent, it will never complete
because the server keeps sending chunks. Netty provides an idle
timer handler that we'll add and configure it to emit an event with
the configIdleRedialTimeout. If the event fires, the endpoint
disconnects itself. Now two things happen individually:

 1) The ClusterManagerBucketRefresher will automatically try to
    get a new streaming connection going when the underlying one
    dies.
 2) The idle time cleanup logic in the pooled service will remove
    the disconnected endpoint from its managed endpoints list making
    sure it does not "leak" around for long.

Change-Id: I0583f33c16335dffee8ad4db28b87749fcfef9cc
Reviewed-on: http://review.couchbase.org/119903
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Graham Pople <grahampople@gmail.com>