gardening: make config output terse
This changeset mainly modifies the memcached bucket config so that
it does not dump the full node info for each hash, since it is
repetitive, verbose and not bringing any value. Instead now it only
dumps each node participating in the ring once.

The couchbase bucket config and the node config has been slightly
polished as well.

Change-Id: I909c254630d8e8e69acb14b76f175a8d48cda849
Reviewed-on: http://review.couchbase.org/119931
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Graham Pople <grahampople@gmail.com>