JVMCBC-793: Use the correct hostname for DNS SRV bootstrap
This changeset first fixes an issue where DNS SRV would not work because
the hostname is passed in as the DNS server instead of the dns hostname.
This obviously prevents DNS SRV from working.

Aside from this, this changeset improves the experience by raising all
kinds of events, including errors as well as regular bootstrap procedure
events.

Finally, the code now also tries to be more intelligent in figuring out
the right hostname for the SRV record.

Change-Id: I863ed61e5bab490784a5795c959dbda785b43b70
Reviewed-on: http://review.couchbase.org/119942
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Graham Pople <grahampople@gmail.com>