JVMCBC-795: Fail requests if service not available
If based on the config we detect that a service is not available at
all (could be either because it wasn't provisioned in the first place
or a failover/rebalance out caused that) sending the request around
for longer will very likely not fix the situation (since someone needs
to come in and re-add that node with the service). It makes more sense
to fail the request with a descriptive exception.

This behavior is now consistent with SDK2.

Change-Id: I6049b5f11fe848dd4ce0abdcf38f6d6659c3ed27
Reviewed-on: http://review.couchbase.org/119956
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Graham Pople <grahampople@gmail.com>