JCBC-1554: Add toBytes to JsonArray and JsonObject
The code already has toString for both of them, but adding a
toBytes allows to directly encode them into their low level
representation without having to go through string (and a potential
byte copy to and from).

Note that there is already the equivalent "from" factory method
for both string and byte array so now they are aligned on the output
side as well.

Change-Id: I89360348a9d0371c80b8d6e7fc115dd20f2b2b24
Reviewed-on: http://review.couchbase.org/119960
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Graham Pople <grahampople@gmail.com>